moved mining drill
